move styles to their own folder
