+++ src/pages/home/styles/index.ts
@@ -1,0 +1,1 @@
+export { default } from './styles.module.css';
